BugFix: File import 0 Bytes.
When importing files this error occurs: Warning: strlen() expects parameter 1 to be string, object given.
Probably when changed to the new php-http/guzzle6-adapter 2.0 probably forgot that the outcome is object instead of a string. So new files will be created in the file system and have no contents eg. 0 Bytes.
FYI: For the image import around line 3096 the change was made this is correct implemented.